+++ src/app/[locale]/(unauth)/landingpage/layout.tsx
@@ -1,4 +1,3 @@
-// import { useTranslations } from 'next-intl';
 import type { ReactNode } from 'react';
 
 import { BaseTemplate } from '@/templates/BaseTemplate';
@@ -8,8 +7,6 @@ export default function Layout({
 }: {
   children: ReactNode;
 }) {
-  // const t = useTranslations('RootLayout');
-
   return (
     <BaseTemplate>
       <div className="w-screen pb-8 pt-16 text-xl sm:pt-24 [&_ul]:my-6">